// src/types/ETodo.ts
export enum ACTIONS {
  FETCH_ALL_TODOS = "fetchAllToDos",
  ADD_TO_DO = "addTodo",
  REMOVE_TODO = "removeToDo",
  UPDATE_TODO = "updateTodo",
}

export enum MUTATIONS {
  ADD_TO_DOS = "addToDos",
  REMOVE_TODO = "removeTodo",
  MUTATE_TODO = "mutateTodo",
  ADD_TO_DO = "addToDo",
  SET_LOADING = "setLoading",
}

export enum GETTERS {
  ALL_TO_DOS = "allTodos",
  LOADING = "loading",
}